fix checkpointing issue
