+++ toolkits/model_checkpoints_convertor/deepseek/hf2mcore_deepseek_v2_moe.py
@@ -24,35 +24,17 @@ sys.path.append(os.path.join(path_dir, "examples"))
 from deepseek_v2.pretrain_deepseek import model_provider
 from megatron_patch.arguments import get_patch_args
 
+from toolkits.model_checkpoints_convertor.utils import (
+    save_state_dict,
+    save_hfmodel
+)
+
 torch.backends.cudnn.deterministic = True
 torch.backends.cudnn.benchmark = False
 torch.backends.cuda.matmul.allow_tf32 = False
 torch.backends.cudnn.allow_tf32 = False
 torch.backends.cuda.enable_mem_efficient_sdp(False)
 torch.backends.cuda.enable_flash_sdp(False)
-
-import numpy as np
-from collections.abc import Mapping, Sequence
-@torch.inference_mode()
-def clone_state_dict(elem):
-    """clone all tensors in the elem to cpu device.
-    """
-    elem_type = type(elem)
-    if isinstance(elem, torch.Tensor):
-        elem = elem.clone()
-    elif isinstance(elem, (np.ndarray, str)):
-        pass
-    elif isinstance(elem, Mapping):
-        elem = dict(elem)
-        for k, v in elem.items():
-            elem[k] = clone_state_dict(v)
-        elem = elem_type(elem)
-    elif isinstance(elem, Sequence):
-        elem = list(elem)
-        for i in range(len(elem)):
-            elem[i] = clone_state_dict(elem[i])
-        elem = elem_type(elem)
-    return elem
 
 def add_model_args(parser):
 
@@ -99,15 +81,6 @@ def name_to_expert_rank(key):
     return expert_rank
 
 def load_megatron_model(args):
-    args.tensor_model_parallel_size = args.target_tensor_model_parallel_size
-    args.pipeline_model_parallel_size = args.target_pipeline_model_parallel_size
-
-    if args.num_experts is not None:
-        args.expert_model_parallel_size = args.target_expert_model_parallel_size
-
-    if args.tensor_model_parallel_size > 1:
-        args.sequence_parallel = True
-
     os.makedirs(args.save, exist_ok=True)
     os.system("cp -rf " + args.hf_ckpt_path + "/config*.json " + args.save)
     os.system("cp -rf " + args.hf_ckpt_path + "/tokenizer* " + args.save)
@@ -118,6 +91,15 @@ def load_megatron_model(args):
     os.system("cp -rf " + args.hf_ckpt_path + "/*.py " + args.load)
 
     model = model_provider()
+
+    args.tensor_model_parallel_size = args.target_tensor_model_parallel_size
+    args.pipeline_model_parallel_size = args.target_pipeline_model_parallel_size
+
+    if args.num_experts is not None:
+        args.expert_model_parallel_size = args.target_expert_model_parallel_size
+
+    if args.tensor_model_parallel_size > 1:
+        args.sequence_parallel = True
 
     model_path = args.load
     tracker_filename = get_checkpoint_tracker_filename(model_path)
@@ -192,7 +174,7 @@ def load_megatron_model(args):
                             pattern = re.compile(r'\d+')
                             res = pattern.findall(k)
                             tgt = re.sub(r"decoder.layers.\d+", "decoder.layers." + str(layers_to_copy[(pp_rank, int(res[0]))]), k)
-                            if 'linear_proj' in k or 'linear_q_proj' or 'linear_q_down_proj' or 'linear_q_up_proj'in k or \
+                            if 'linear_proj' in k or 'linear_q_proj' in k or 'linear_q_down_proj' in k or 'linear_q_up_proj'in k or \
                                     'linear_kv_up_proj' in k or 'decoder.layers.0.mlp.linear_fc2' in k or \
                                     'decoder.layers.0.mlp.linear_fc1' in k or 'shared_experts.linear_fc1' in k or 'shared_experts.linear_fc2' in k:
                                 if ep_rank ==0:
@@ -352,16 +334,6 @@ def convert_checkpoint_from_transformers_to_megatron(hfmodel, mgmodel, args):
         if args.untie_embeddings_and_output_weights:
             mgmodel.output_layer.weight.copy_(hfmodel.lm_head.weight)
 
-
-def save_state_dict(args, model, checkpoint_name):
-    state_dict = {}
-    state_dict['args'] = args
-    state_dict['checkpoint_version'] = 3.0
-    state_dict['iteration'] = 0
-    state_dict['model'] = model
-    os.makedirs(os.path.dirname(checkpoint_name), exist_ok=True)
-    print(f'save model part {checkpoint_name}')
-    torch.save(clone_state_dict(state_dict), checkpoint_name)
 
 def check_layer(layers_to_copy, k):
     pattern = re.compile(r"decoder.layers.\d+")
@@ -592,41 +564,6 @@ def save_mgmodel(mgmodel, args):
 
     print(f'megatron model is save to {args.save}')
 
-
-def save_hfmodel(args, model):
-    output_state_dict = model.state_dict()
-    max_shard_size = "10GB"
-    shards, index = shard_checkpoint(output_state_dict, max_shard_size=max_shard_size)
-    os.makedirs(args.save, exist_ok=True)
-    for shard_file, shard in shards.items():
-        if args.save_safetensors:
-            shard_file = shard_file.replace("pytorch_", "")
-            shard_file = shard_file.replace(".bin", ".safetensors")
-            target_file = os.path.join(args.save, shard_file)
-            print(f'huggingface model is save to {target_file}')
-            new_shard = {}
-            for k, v in shard.items():
-                new_shard[k] = copy.deepcopy(v)
-            safetensors.torch.save_file(clone_state_dict(new_shard), target_file, metadata={"format": "pt"})
-        else:
-            target_file = os.path.join(args.save, shard_file)
-            print(f'huggingface model is save to {target_file}')
-            torch.save(clone_state_dict(shard), target_file)
-
-    if index is None:
-        print(f"Model weights saved in {os.path.join(args.save, WEIGHTS_NAME)}")
-    else:
-        save_index_file = os.path.join(args.save, WEIGHTS_INDEX_NAME)
-        # Save the index as well
-        with open(save_index_file, "w", encoding="utf-8") as f:
-            content = json.dumps(index, indent=2, sort_keys=True) + "\n"
-            f.write(content)
-        print(
-            f"The model is bigger than the maximum size per checkpoint ({max_shard_size}) and is going to be "
-            f"split in {len(shards)} checkpoint shards. You can find where each parameters has been saved in the "
-            f"index located at {save_index_file}."
-        )
-
 def check_hf_mg_forward(hfmodel, mgmodel, mgargs):
     hf_hiddens = [{} for _ in range(mgargs.num_layers)]
     mg_hiddens = [{} for _ in range(mgargs.num_layers)]
@@ -844,7 +781,6 @@ def check_hf_mg_forward(hfmodel, mgmodel, mgargs):
         diff_num = ((hflogits - mglogits) > epsilon).sum()
         diff_max = (hflogits - mglogits).abs().max()
         print(f'logits: {same_num}, diff>{epsilon}:[{diff_num}/{hflogits.numel()}] diff_max:{diff_max}')
-
 
 def add_extra_args(parser):
     parser = get_patch_args(parser)
